# docs/swagger.json
{
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/cycles": {
            "get": {
                "description": "Get all cycles that are available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mysql"
                ],
                "summary": "Get all cycles",
                "operationId": "get-all-cycles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.Cycle"
                        }
                    }
                }
            }
        },
        "/cycles/{cycle_id}/mysql/engine-innodb-status": {
            "get": {
                "description": "Get a list of mysql engine innoDB statuses belonging to a cycle id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mysql"
                ],
                "summary": "Get all mysql engine InnoDB status by cycle id",
                "operationId": "get-all-mysql-engine-innodb",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.SlaveStatus"
                        }
                    }
                }
            }
        },
        "/cycles/{cycle_id}/mysql/processlist": {
            "get": {
                "description": "Get a list of mysql processlist belonging to a cycle id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mysql"
                ],
                "summary": "Get all mysql processlist by cycle id",
                "operationId": "get-all-mysql-process-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.MysqlProcessList"
                        }
                    }
                }
            }
        },
        "/cycles/{cycle_id}/mysql/slave-status": {
            "get": {
                "description": "Get a list of mysql slave status belonging to a cycle id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mysql"
                ],
                "summary": "Get all mysql slave statuses by cycle id",
                "operationId": "get-all-mysql-slave-status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.SlaveStatus"
                        }
                    }
                }
            }
        },
        "/cycles/{cycle_id}/mysql/threads": {
            "get": {
                "description": "Get a list of mysql threads belonging to a cycle id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mysql"
                ],
                "summary": "Get all mysql threads by cycle id",
                "operationId": "get-all-mysql-threads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.Thread"
                        }
                    }
                }
            }
        },
        "/cycles/{cycle_id}/system/process-list": {
            "get": {
                "description": "Get a list of system processlist belonging to a cycle id, this is always 1 entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get all system processlist by cycle id, this is always 1 entry",
                "operationId": "get-all-system-process-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.UnixProcessList"
                        }
                    }
                }
            }
        },
        "/cycles/{cycle_id}/system/top": {
            "get": {
                "description": "Get a list of system top belonging to a cycle id, this is always 1 entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get all system top by cycle id, this is always 1 entry",
                "operationId": "get-all-system-top",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "cycle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/data.Top"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.Cycle": {
            "type": "object",
            "properties": {
                "cycle": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "ran_at": {
                    "type": "string"
                },
                "took": {
                    "type": "string"
                }
            }
        },
        "data.MysqlProcessList": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "db": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "rows_examined": {
                    "type": "integer"
                },
                "rows_sent": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "data.SlaveStatus": {
            "type": "object",
            "properties": {
                "auto_position": {
                    "type": "integer"
                },
                "channel_name": {
                    "type": "string"
                },
                "connect_retry": {
                    "type": "string"
                },
                "exec_master_log_pos": {
                    "type": "integer"
                },
                "executed_gtid_set": {
                    "type": "string"
                },
                "last_errno": {
                    "type": "integer"
                },
                "last_error": {
                    "type": "string"
                },
                "last_io_errno": {
                    "type": "integer"
                },
                "last_io_error": {
                    "type": "string"
                },
                "last_io_error_timestamp": {
                    "type": "string"
                },
                "last_sql_errno": {
                    "type": "integer"
                },
                "last_sql_error": {
                    "type": "string"
                },
                "last_sql_error_timestamp": {
                    "type": "string"
                },
                "master_bind": {
                    "type": "string"
                },
                "master_host": {
                    "type": "string"
                },
                "master_info_file": {
                    "type": "string"
                },
                "master_log_file": {
                    "type": "string"
                },
                "master_port": {
                    "type": "string"
                },
                "master_retry_count": {
                    "type": "integer"
                },
                "master_server_id": {
                    "type": "integer"
                },
                "master_ssl_allowed": {
                    "type": "string"
                },
                "master_ssl_ca_file": {
                    "type": "string"
                },
                "master_ssl_ca_path": {
                    "type": "string"
                },
                "master_ssl_cert": {
                    "type": "string"
                },
                "master_ssl_cipher": {
                    "type": "string"
                },
                "master_ssl_crl": {
                    "type": "string"
                },
                "master_ssl_crlpath": {
                    "type": "string"
                },
                "master_ssl_key": {
                    "type": "string"
                },
                "master_ssl_verify_server_cert": {
                    "type": "string"
                },
                "master_tls_version": {
                    "type": "string"
                },
                "master_user": {
                    "type": "string"
                },
                "master_uuid": {
                    "type": "string"
                },
                "read_master_log_pos": {
                    "type": "integer"
                },
                "relay_log_file": {
                    "type": "string"
                },
                "relay_log_pos": {
                    "type": "integer"
                },
                "relay_log_space": {
                    "type": "integer"
                },
                "relay_master_log_file": {
                    "type": "string"
                },
                "replicate_do_db": {
                    "type": "string"
                },
                "replicate_do_table": {
                    "type": "string"
                },
                "replicate_ignore_db": {
                    "type": "string"
                },
                "replicate_ignore_server_ids": {
                    "type": "integer"
                },
                "replicate_ignore_table": {
                    "type": "string"
                },
                "replicate_rewrite_db": {
                    "type": "string"
                },
                "replicate_wild_do_table": {
                    "type": "string"
                },
                "replicate_wild_ignore_table": {
                    "type": "string"
                },
                "retrieved_gtid_set": {
                    "type": "string"
                },
                "seconds_behind_master": {
                    "type": "integer"
                },
                "skip_counter": {
                    "type": "integer"
                },
                "slave_io_running": {
                    "type": "string"
                },
                "slave_io_state": {
                    "type": "string"
                },
                "slave_io_status": {
                    "type": "string"
                },
                "slave_sql_running": {
                    "type": "string"
                },
                "slave_sql_running_state": {
                    "type": "string"
                },
                "slave_status": {
                    "type": "string"
                },
                "sql_delay": {
                    "type": "integer"
                },
                "sql_remaining_delay": {
                    "type": "string"
                },
                "until_condition": {
                    "type": "string"
                },
                "until_log_file": {
                    "type": "string"
                },
                "until_log_pos": {
                    "type": "integer"
                }
            }
        },
        "data.Thread": {
            "type": "object",
            "properties": {
                "connection_type": {
                    "type": "integer"
                },
                "history": {
                    "type": "string"
                },
                "instrumented": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_thread_id": {
                    "type": "integer"
                },
                "process_list_command": {
                    "type": "string"
                },
                "process_list_db": {
                    "type": "string"
                },
                "process_list_host": {
                    "type": "string"
                },
                "process_list_id": {
                    "type": "integer"
                },
                "process_list_info": {
                    "type": "string"
                },
                "process_list_state": {
                    "type": "string"
                },
                "process_list_time": {
                    "type": "string"
                },
                "process_list_user": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "integer"
                },
                "thread_os_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "data.Top": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "data.UnixProcessList": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        }
    },
    "swagger": "2.0"
}